// src/enums/enums.ts
export enum TipoAeronave {
  COMERCIAL = "Comercial",
  MILITAR = "Militar"
}

export enum TipoPeca {
  NACIONAL = "Nacional",
  IMPORTADA = "Importada"
}

export enum StatusPeca {
  EM_PRODUCAO = "Em produção",
  EM_TRANSPORTE = "Em transporte",
  PRONTA = "Pronta"
}

export enum StatusEtapa {
  PENDENTE = "Pendente",
  ANDAMENTO = "Em andamento",
  CONCLUIDA = "Concluída"
}

export enum NivelPermissao {
  ADMINISTRADOR = "Administrador",
  ENGENHEIRO = "Engenheiro",
  OPERADOR = "Operador"
}

export enum TipoTeste {
  ELETRICO = "Elétrico",
  HIDRAULICO = "Hidráulico",
  AERODINAMICO = "Aerodinâmico"
}

export enum ResultadoTeste {
  APROVADO = "Aprovado",
  REPROVADO = "Reprovado"
}